feat: 지도 앱을 이동 함수 util로 분리하고 연결시키기
